Island: Move AWSService to service/aws/
