+++ monkey/monkey_island/cc/services/__init__.py
@@ -4,4 +4,8 @@ from .directory_file_storage_service import DirectoryFileStorageService
 from .authentication.authentication_service import AuthenticationService
 from .authentication.json_file_user_datastore import JsonFileUserDatastore
 
-from .aws_service import AWSService
+from .aws import AWSService
+
+# TODO: This is a temporary import to keep some tests passing. Remove it before merging #1928 to
+#       develop.
+from .aws import aws_service
