+++ monkey/monkey_island/cc/services/aws/__init__.py
@@ -1,0 +1,1 @@
+from .aws_service import AWSService
