+++ monkey/monkey_island/cc/services/aws/aws_service.py
